Fix cycle/last dtype in cat240toHDF5.py
cycle/last used NumPy's default integer dtype, which made HDF5
store it as int64 while cycle/first was uint32. Build cycle/last
with cycle/first's dtype so both cycle indices use uint32.

+++ examples/cat240toHDF5.py
@@ -89,7 +89,7 @@ def _find_cycle_start(az, *, verbose=False):
 
 def _make_cycle_lookup(blocks, *, verbose=False):
     first = _find_cycle_start(blocks["az1"], verbose=verbose)
-    last = np.arange(first.size)
+    last = np.arange(first.size, dtype=first.dtype)
 
     t = blocks["tod"]
     mask = np.full(t.shape, True, dtype=np.bool)
